Split service binaries to separate crates
COPR Module Manager: managerforlustre/django-bump
COPR Module Manager: managerforlustre/service-bump

The files under iml-services are structured as a single crate.

However, we produce a number of binaries from this one crate.

Split the iml-service files into separate crates, this will allow us to
have separate dependencies per crate.
