--- iml-services/Cargo.toml
@@ -1,34 +1,0 @@
-[package]
-name = "iml-services"
-version = "0.1.0"
-authors = ["IML Team <iml@whamcloud.com>"]
-edition = "2018"
-
-[dependencies]
-lapin-futures = "0.18.0"
-futures = "0.1"
-tokio = "0.1.21"
-tokio-timer = "0.2.11"
-env_logger = "0.7"
-log = "0.4"
-serde = { version = "1", features = ["derive"] }
-serde_json = "1.0"
-failure = "0.1.5"
-warp = "0.1.16" 
-iml-wire-types = { path = "../iml-wire-types", version = "0.2" }
-iml-rabbit = { path = "../iml-rabbit", version = "0.1.0" }
-iml-manager-env = { path = "../iml-manager-env", version = "0.1.0" }
-tokio-runtime-shutdown = { path = "../tokio-runtime-shutdown", version = "0.1.0" }
-parking_lot = "0.9"
-
-[dev-dependencies]
-iml-agent-comms = { path = "../iml-agent-comms", version = "0.1.0"}
-rand = "0.7"
-
-[[bin]]
-name = "iml-stratagem"
-path = "src/services/stratagem.rs"
-
-[[bin]] 
-name = "iml-action-runner"
-path = "src/services/action_runner/main.rs"
