+++ iml-services/iml-action-runner/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "iml-action-runner"
+version = "0.1.0"
+authors = ["IML Team <iml@whamcloud.com>"]
+edition = "2018"
+
+[dependencies]
+futures = "0.1.29"
+tokio = "0.1.22"
+tokio-timer = "0.2.11"
+env_logger = "0.7"
+log = "0.4.8"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1.0"
+failure = "0.1.5"
+warp = "0.1.20" 
+iml-wire-types = { path = "../../iml-wire-types", version = "0.2" }
+iml-rabbit = { path = "../../iml-rabbit", version = "0.1.0" }
+iml-manager-env = { path = "../../iml-manager-env", version = "0.1.0" }
+iml-service-queue = { path = "../iml-service-queue", version = "0.1.0" }
+tokio-runtime-shutdown = { path = "../../tokio-runtime-shutdown", version = "0.1.0" }
+parking_lot = "0.9"
+
+[dev-dependencies]
+iml-agent-comms = { path = "../../iml-agent-comms", version = "0.1.0"}
+rand = "0.7.2"
+lapin-futures = "0.18.0"
