+++ iml-services/iml-action-runner/src/lib.rs
@@ -2,5 +2,7 @@
 // Use of this source code is governed by a MIT-style
 // license that can be found in the LICENSE file.
 
-pub mod service_queue;
-pub mod services;
+pub mod data;
+pub mod error;
+pub mod receiver;
+pub mod sender;
