+++ iml-services/iml-service-queue/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "iml-service-queue"
+version = "0.1.0"
+authors = ["IML Team <iml@whamcloud.com>"]
+edition = "2018"
+
+[dependencies]
+lapin-futures = "0.18.0"
+tokio = "0.1.21"
+log = "0.4.6"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1.0"
+failure = "0.1.5"
+iml-wire-types = { path = "../../iml-wire-types", version = "0.2" }
+iml-rabbit = { path = "../../iml-rabbit", version = "0.1.0" }
