+++ iml-services/iml-service-queue/src/lib.rs
@@ -1,0 +1,5 @@
+// Copyright (c) 2019 DDN. All rights reserved.
+// Use of this source code is governed by a MIT-style
+// license that can be found in the LICENSE file.
+
+pub mod service_queue;
