+++ iml-services/iml-stratagem/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "iml-stratagem"
+version = "0.1.0"
+authors = ["IML Team <iml@whamcloud.com>"]
+edition = "2018"
+
+[dependencies]
+futures = "0.1"
+tokio = "0.1"
+env_logger = "0.7"
+log = "0.4.6"
+iml-wire-types = { path = "../../iml-wire-types", version = "0.2" }
+iml-rabbit = { path = "../../iml-rabbit", version = "0.1.0" }
+iml-service-queue = { path = "../iml-service-queue", version = "0.1.0" }
